pyproject.toml
[tool.poetry]
name = "wagers-dev"
version = "0.1.0"
description = ""
authors = ["Loopeq <arsen.00z11@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.3"
webshareproxy = "^0.0.10"
pydantic-settings = "^2.4.0"
pytest = "^8.3.2"
fake-useragent = "^1.5.1"
aiohttp = "^3.10.5"
aiocache = "^0.12.2"
apscheduler = "^3.10.4"
sqlalchemy = "^2.0.32"
asyncpg = "^0.29.0"
alembic = "^1.13.2"
fastapi = "^0.114.2"
uvicorn = {extras = ["standard"], version = "^0.30.6"}
websockets = "^13.1"
colorlog = "^6.8.2"
gunicorn = "^23.0.0"
cachetools = "^5.5.0"
passlib = "^1.7.4"
pyjwt = "^2.10.1"
python-multipart = "^0.0.20"
pytz = "^2024.2"
python-json-logger = "^3.2.1"
blinker = "1.7.0"
pytest-playwright = "^0.7.0"
beautifulsoup4 = "^4.13.4"
rapidfuzz = "^3.13.0"
g4f = {extras = ["all"], version = "^0.6.0.3"}
aiogram = "^3.22.0"
watchdog = "^6.0.0"
arq = "^0.27.0"
aioredis = "^2.0.1"
pika = "^1.3.2"
aio-pika = "^9.6.2"

[tool.poetry.group.dev.dependencies]
openpyxl = "^3.1.5"
flake8 = "^7.1.1"
black = "^25.1.0"
ruff = "^0.12.9"
isort = "^8.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
server = 'src.main:main'
